+++ packages/core/src/components/async-server/async-server.hook-builder.ts
@@ -133,7 +133,7 @@ export class AsyncServerHookBuilder<TPayload = unknown> extends BaseHookBuilder 
 			params: {
 				schema: schema ?? undefined,
 				lookupKey: stepParams.messageType ?? stepParams.eventType,
-				lookupDirection: "clientMessage",
+				lookupDirection: stepParams.eventType ? "serverMessage" : "clientMessage",
 			},
 		});
 	}